[#133][MODIFY] refactoring package
