--- src/main/java/com/mashup/thing/endpage/dto/ResEndPageReviewDto.java
@@ -1,19 +1,0 @@
-package com.mashup.thing.endpage.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-public class ResEndPageReviewDto {
-
-    private Long id;
-    private LocalDateTime createAt;
-    private String nickName;
-    private String profileUrl;
-    private String text;
-    private String liked;
-    private Boolean owner;
-}
