--- src/main/java/com/mashup/thing/endpage/dto/ResEndPageVideoDto.java
@@ -1,18 +1,0 @@
-package com.mashup.thing.endpage.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-public class ResEndPageVideoDto {
-
-    private Long id;
-    private String thumbnail;
-    private String youtubeVideoId;
-    private LocalDateTime publishedAt;
-    private String title;
-
-}
