--- src/main/java/com/mashup/thing/ranking/dto/ResRankingDto.java
@@ -1,18 +1,0 @@
-package com.mashup.thing.ranking.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ResRankingDto {
-
-    private Long id;
-    private String name;
-    private Long ranking;
-    private String viewCount;
-    private String subscriberCount;
-    private String thumbnail;
-    private String bannerImgUrl;
-    private String soaring;
-}
