--- src/main/java/com/mashup/thing/recommendation/dto/ResVideoDto.java
@@ -1,13 +1,0 @@
-package com.mashup.thing.recommendation.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ResVideoDto {
-    private Long id;
-    private String thumbnail;
-    private String youtubeVideoId;
-    private String title;
-}
