+++ src/main/java/com/mashup/thing/tag/TagService.java
@@ -1,7 +1,6 @@
 package com.mashup.thing.tag;
 
 import com.mashup.thing.tag.domain.Tag;
-import com.mashup.thing.tag.dto.ResTagsDto;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
