--- src/main/java/com/mashup/thing/tag/dto/ResTagDto.java
@@ -1,18 +1,0 @@
-package com.mashup.thing.tag.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@Builder
-public class ResTagDto {
-
-    private String category;
-    private List<String> list;
-
-    public void addSubTag(String subTag) {
-        this.list.add(subTag);
-    }
-}
