--- src/main/java/com/mashup/thing/tag/dto/ResTagsDto.java
@@ -1,13 +1,0 @@
-package com.mashup.thing.tag.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@Builder
-public class ResTagsDto {
-
-    private List<ResTagDto> tags;
-}
