--- src/main/java/com/mashup/thing/user/dto/ResCategoryDto.java
@@ -1,12 +1,0 @@
-package com.mashup.thing.user.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ResCategoryDto {
-    private Long id;
-    private String categoryType;
-
-}
