+++ src/main/java/com/mashup/thing/youtuber/dto/ReqRecommendationDto.java
@@ -1,4 +1,4 @@
-package com.mashup.thing.recommendation.dto;
+package com.mashup.thing.youtuber.dto;
 
 import lombok.Getter;
 import lombok.Setter;
