+++ src/main/java/com/mashup/thing/youtuber/dto/ResHomeDto.java
@@ -1,5 +1,6 @@
-package com.mashup.thing.recommendation.dto;
+package com.mashup.thing.youtuber.dto;
 
+import lombok.Builder;
 import lombok.Getter;
 
 import java.util.ArrayList;
@@ -9,6 +10,19 @@ import java.util.List;
 public class ResHomeDto {
     private List<ResYouTuberDto> resYouTuberDtos;
     private List<ResSoaringYouTuberDto> resSoaringYouTuberDtos;
+
+    @Getter
+    @Builder
+    public static class ResSoaringYouTuberDto {
+
+        private Long id;
+        private String name;
+        private String thumbnail;
+        private List<String> tag;
+        private Double soaring;
+
+    }
+
 
     public ResHomeDto(List<ResYouTuberDto> resYouTuberDtos, List<ResSoaringYouTuberDto> resSoaringYouTuberDtos) {
         this.resYouTuberDtos = resYouTuberDtos;
