+++ src/main/java/com/mashup/thing/youtuber/dto/ResRecommendationDto.java
@@ -1,4 +1,4 @@
-package com.mashup.thing.recommendation.dto;
+package com.mashup.thing.youtuber.dto;
 
 import lombok.Getter;
